+++ app/src/main/java/com/fibelatti/photowidget/model/PhotoWidgetColors.kt
@@ -11,4 +11,20 @@ data class PhotoWidgetColors(
     val opacity: Float = DEFAULT_OPACITY,
     val saturation: Float = DEFAULT_SATURATION,
     val brightness: Float = DEFAULT_BRIGHTNESS,
-) : Parcelable
+) : Parcelable {
+
+    companion object {
+
+        /**
+         * Saturation is persisted using a [0..200] range, while the UI expects a [-100..100]
+         * range. This function converts the persisted value to the UI value.
+         */
+        fun pickerSaturation(saturation: Float): Float = saturation - 100
+
+        /**
+         * The saturation picker UI uses a [-100..100] range, while the persisted value uses
+         * [0..200]. This function converts the UI value to the persisted value.
+         */
+        fun persistenceSaturation(saturation: Float): Float = saturation + 100
+    }
+}